pyproject.toml
[tool.poetry]
name = "LLM Examples"
version = "1.0.0"
description = "LLM Examples"
authors = ["jggomez <juan.gomez01@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^1.0.1"
rich = "^13.7.1"
pillow = "^10.3.0"
opencv-python = "^4.9.0.80"
sentence-transformers = "^2.7.0"
pypdf2 = "^3.0.1"
langchain-google-genai = "^1.0.3"
google-cloud-firestore = "^2.16.0"
numpy = "^1.26.4"
langchain-google-vertexai = "^1.0.5"
google-generativeai = "^0.7.1"
google-cloud-aiplatform = "^1.57.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
